requirements.txt
numpy==1.16.4
scipy==1.3.0
tensorflow==1.14.0
Pillow==6.0.0
scikit-learn==0.21.2
